+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # dev
 
+* decide whether you want to use local bower/grunt or the global version (saves up to 55MB!)
 * made activation of SVG Sprites configurable
 * load `svg-sprite.svg` from svgstore with ajax #13
 * made the `--target` of `grunt build` and `grunt dist` configurable #10